+++ .env.example
@@ -1,0 +1,3 @@
+
+VITE_MAPBOX_PUBLIC_KEY=""
+VITE_SYNOPTIC_KEY=""
