+++ README.md
@@ -1,14 +1,21 @@
 # Mausam
 
-A data-vis solution for meteorological data.
+Mausam is a weather application that displays weather data on a map. The application uses the [Synoptic Data](https://synopticdata.com/) API to fetch weather data and [Mapbox](https://www.mapbox.com/) for the map.
+
+## Getting Started
+
+- Get api keys from [Mapbox](https://www.mapbox.com/) and [Synoptic Data](https://customer.synopticdata.com/credentials/).
+- Rename `.env.example` to `.env` and add the api keys.
+- Run `npm install` to install the dependencies.
+- Run `npm run dev` to start the development server.
+
 
 ## Built With
 
 - React
 - Tailwind
-- Deck.gl / React-map-gl
+- React-map-gl
 
-## Prerequisites
 
 ### Implementation Notes
 
@@ -25,7 +32,6 @@ return (
     <Map id="mapInstanceA" />
     <Map id="mapInstanceB" />
 )
-
 ```
 
 #### All the dates that are not unit timestamp follow [ISO 8601 format](https://en.wikipedia.org/wiki/ISO_8601).
@@ -35,6 +41,3 @@ return (
 - [Mapbox](https://www.mapbox.com/)
 - [Synoptic Data](https://customer.synopticdata.com/credentials/)
 
-### Resources
-
-- [CartoDB/carto-react-template/](https://github.com/CartoDB/carto-react-template/tree/7c6200f9bedafba6e1230360a7e4e56e2bc9eb70)
